+++ scripts/macros/classFeatures/artificer/infusions/homunculusServant.js
@@ -24,6 +24,8 @@ async function item({speaker, actor, token, character, item, args, scope, workfl
     let hpValue = await new Roll(hpFormula).evaluate({async: true});
     let name = chris.getConfiguration(workflow.item, 'name') ?? 'Homunculus Servant';
     if (name === '') name = 'Homunculus Servant';
+    let meleeAttackBonus = await new Roll(workflow.actor.system.bonuses.msak.attack + ' + 0').roll({async: true});
+    let rangedAttackBonus = await new Roll(workflow.actor.system.bonuses.rsak.attack + ' + 0').roll({async: true});
     let updates = {
         'actor': {
             'name': name,
@@ -47,8 +49,8 @@ async function item({speaker, actor, token, character, item, args, scope, workfl
                 'chris-premades': {
                     'summon': {
                         'attackBonus': {
-                            'melee': chris.getSpellMod(workflow.item) - sourceActor.system.abilities.int.mod + Number(workflow.actor.system.bonuses.msak.attack),
-                            'ranged': chris.getSpellMod(workflow.item) - sourceActor.system.abilities.int.mod + Number(workflow.actor.system.bonuses.rsak.attack)
+                            'melee': chris.getSpellMod(workflow.item) - sourceActor.system.abilities.int.mod + meleeAttackBonus.total,
+                            'ranged': chris.getSpellMod(workflow.item) - sourceActor.system.abilities.int.mod + rangedAttackBonus.total
                         }
                     }
                 }