+++ scripts/macros/classFeatures/barbarian/giant/elementalCleaver.js
@@ -79,7 +79,6 @@ async function mutate(workflow, selected) {
 async function item({speaker, actor, token, character, item, args, scope, workflow}) {
     if (!workflow.actor || !workflow.token) return;
     let weapons = workflow.actor.items.filter(i => i.type === 'weapon' && i.system.equipped);
-    console.log(weapons);
     if (!weapons.length) return;
     let selected;
     if (weapons.length === 1) {