+++ scripts/macros/classFeatures/bard/collegeOfCreation/animatingPerformance.js
@@ -42,6 +42,8 @@ async function item({speaker, actor, token, character, item, args, scope, workfl
     }
     let name = chris.getConfiguration(workflow.item, 'name') ?? 'Dancing Item';
     if (name === '') name = 'Dancing Item';
+    let meleeAttackBonus = await new Roll(workflow.actor.system.bonuses.msak.attack + ' + 0').roll({async: true});
+    let rangedAttackBonus = await new Roll(workflow.actor.system.bonuses.rsak.attack + ' + 0').roll({async: true});
     let updates = {
         'actor': {
             'name': name,
@@ -65,8 +67,8 @@ async function item({speaker, actor, token, character, item, args, scope, workfl
                 'chris-premades': {
                     'summon': {
                         'attackBonus': {
-                            'melee': chris.getSpellMod(workflow.item) - sourceActor.system.abilities.str.mod + Number(workflow.actor.system.bonuses.msak.attack),
-                            'ranged': chris.getSpellMod(workflow.item) - sourceActor.system.abilities.str.mod + Number(workflow.actor.system.bonuses.rsak.attack)
+                            'melee': chris.getSpellMod(workflow.item) - sourceActor.system.abilities.str.mod + meleeAttackBonus.total,
+                            'ranged': chris.getSpellMod(workflow.item) - sourceActor.system.abilities.str.mod + rangedAttackBonus.total
                         }
                     }
                 }